backend/requirements.txt
# ═══════════════════════════════════════════════════════════════
# NeuralMigrate — Backend Dependencies
# Option A setup: Groq (free LLM) + HuggingFace (free embeddings)
# No Docker, no Ollama, no GPU required.
# ═══════════════════════════════════════════════════════════════

# ── Core LangChain / LangGraph ────────────────────────────────
langchain>=0.2.0,<0.4.0
langchain-core>=0.2.0,<0.4.0
langchain-community>=0.2.0,<0.4.0
langchain-openai>=0.1.0,<0.3.0
langchain-groq>=0.1.0,<0.3.0
langgraph>=0.1.0,<0.3.0

# ── LLM — FREE ───────────────────────────────────────────────
groq>=0.9.0,<1.0.0                # Groq free tier (llama3-8b-8192)

# ── Embeddings — FREE, ~90MB, CPU only ───────────────────────
sentence-transformers>=2.7.0,<4.0.0   # all-MiniLM-L6-v2 (384-dim)
# Downloads the model on first run — cached in ~/.cache/huggingface

# ── LLM — optional paid providers ────────────────────────────
openai>=1.30.0,<2.0.0             # OpenAI + DeepSeek (OpenAI-compatible)
# Ollama ships via langchain-community — no extra package needed

# ── Databases ─────────────────────────────────────────────────
neo4j>=5.20.0,<6.0.0              # Neo4j Aura free tier compatible
psycopg2-binary>=2.9.9,<3.0.0    # PostgreSQL (Neon free tier)
pgvector>=0.2.4,<0.4.0            # pgvector Python helpers

# ── AST parsing ───────────────────────────────────────────────
tree-sitter>=0.22.0,<1.0.0


# ── Data validation ───────────────────────────────────────────
pydantic>=2.7.0,<3.0.0

# ── API server ────────────────────────────────────────────────
fastapi>=0.111.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
gunicorn>=22.0.0,<23.0.0          # Production WSGI (Render)
python-multipart>=0.0.9

# ── Utilities ─────────────────────────────────────────────────
numpy>=1.26.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0

# ── Testing ───────────────────────────────────────────────────
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.23.0,<1.0.0
httpx>=0.27.0,<1.0.0
pytest-cov>=5.0.0,<6.0.0
